ECDH-SS support
