+++ tests/Component/Encryption/EncryptionTest.php
@@ -22,6 +22,10 @@ use Jose\Component\Encryption\Algorithm\KeyEncryption\ECDHES;
 use Jose\Component\Encryption\Algorithm\KeyEncryption\ECDHESA128KW;
 use Jose\Component\Encryption\Algorithm\KeyEncryption\ECDHESA192KW;
 use Jose\Component\Encryption\Algorithm\KeyEncryption\ECDHESA256KW;
+use Jose\Component\Encryption\Algorithm\KeyEncryption\ECDHSS;
+use Jose\Component\Encryption\Algorithm\KeyEncryption\ECDHSSA128KW;
+use Jose\Component\Encryption\Algorithm\KeyEncryption\ECDHSSA192KW;
+use Jose\Component\Encryption\Algorithm\KeyEncryption\ECDHSSA256KW;
 use Jose\Component\Encryption\Algorithm\KeyEncryption\PBES2HS256A128KW;
 use Jose\Component\Encryption\Algorithm\KeyEncryption\PBES2HS384A192KW;
 use Jose\Component\Encryption\Algorithm\KeyEncryption\PBES2HS512A256KW;
@@ -77,6 +81,10 @@ abstract class EncryptionTest extends TestCase
             $this->algorithmManagerFactory->add('ECDH-ES+A128KW', new ECDHESA128KW());
             $this->algorithmManagerFactory->add('ECDH-ES+A192KW', new ECDHESA192KW());
             $this->algorithmManagerFactory->add('ECDH-ES+A256KW', new ECDHESA256KW());
+            $this->algorithmManagerFactory->add('ECDH-SS', new ECDHSS());
+            $this->algorithmManagerFactory->add('ECDH-SS+A128KW', new ECDHSSA128KW());
+            $this->algorithmManagerFactory->add('ECDH-SS+A192KW', new ECDHSSA192KW());
+            $this->algorithmManagerFactory->add('ECDH-SS+A256KW', new ECDHSSA256KW());
             $this->algorithmManagerFactory->add('PBES2-HS256+A128KW', new PBES2HS256A128KW());
             $this->algorithmManagerFactory->add('PBES2-HS384+A192KW', new PBES2HS384A192KW());
             $this->algorithmManagerFactory->add('PBES2-HS512+A256KW', new PBES2HS512A256KW());
